fix(mcp): centralize interactive tool filtering
